Move web3 to common
