+++ src/common/services/web3.js
@@ -1,5 +1,12 @@
 import Web3 from 'web3';
 
-const web3 = new Web3();
+const rpcUrl = process.env.RPC_URL || 'http://localhost:8545';
 
-export default web3;
+const provider = new Web3.providers.WebsocketProvider(rpcUrl);
+const web3 = new Web3(provider);
+
+const connect = async () => {
+  return (await web3.eth.getBlock('latest')).number;
+};
+
+export default { connect, web3 };
