+++ src/server/index.js
@@ -12,7 +12,7 @@ import morgan from 'morgan';
 
 dotenv.config();
 
-import blockchain from '~/server/blockchain';
+import blockchain from '~/common/services/web3';
 import db from '~/server/database';
 import errorsMiddleware from '~/server/middlewares/errors';
 import logger from '~/server/helpers/logger';
